Copy past content of DAGNode in Node
