+++ Sofa/framework/Simulation/Graph/CMakeLists.txt
@@ -14,7 +14,7 @@ set(HEADER_FILES
 set(SOURCE_FILES
     ${SOFASIMULATIONGRAPH_SRC}/init.cpp
     ${SOFASIMULATIONGRAPH_SRC}/initSofaSimulationGraph.cpp
-    ${SOFASIMULATIONGRAPH_SRC}/NodeDagImpl.cpp
+    ${SOFASIMULATIONGRAPH_SRC}/DAGNode.cpp
     ${SOFASIMULATIONGRAPH_SRC}/DAGSimulation.cpp
 )
